rm unused interface, move platform -> base (#231306)
