+++ src/vs/workbench/contrib/chat/browser/actions/chatActions.ts
@@ -10,7 +10,6 @@ import { fromNowByDay } from '../../../../../base/common/date.js';
 import { KeyCode, KeyMod } from '../../../../../base/common/keyCodes.js';
 import { DisposableStore } from '../../../../../base/common/lifecycle.js';
 import { ThemeIcon } from '../../../../../base/common/themables.js';
-import { URI } from '../../../../../base/common/uri.js';
 import { ICodeEditor } from '../../../../../editor/browser/editorBrowser.js';
 import { EditorAction2, ServicesAccessor } from '../../../../../editor/browser/editorExtensions.js';
 import { Position } from '../../../../../editor/common/core/position.js';
@@ -63,12 +62,6 @@ export interface IChatViewOpenOptions {
 	 * Whether a screenshot of the focused window should be taken and attached
 	 */
 	attachScreenshot?: boolean;
-}
-
-export interface IChatImageAttachment {
-	id: string;
-	name: string;
-	value: URI | Uint8Array;
 }
 
 export interface IChatViewOpenRequestEntry {
